+++ src/module/enums/ArmorLocationsEnum.ts
@@ -1,3 +1,7 @@
+/**
+ * @deprecated - please use `CONFIG.SWADE.CONST.ARMOR_LOCATIONS` instead
+ * @since v1.1.0
+ */
 export enum ArmorLocation {
   Head = 'head',
   Torso = 'torso',