+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added default duration of 10 rounds to Berserk status effect template
 - Added End of Turn prompt expiration to Berserk status effect template
 - Added Start of Turn auto expiration to Defending status effect template
+- Added `CONFIG.SWADE.CONST` object which will contain constants and enums
 
 ### Changed
 
@@ -38,6 +39,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Started deprecation of `game.swade.SwadeEntityTweaks`. The class has been moved to `game.swade.apps.SwadeEntityTweaks` and the old accessor will be removed with v1.2.0
 - Started deprecation of `game.swade.itemChatCardHelper`. The class has been moved to `game.swade.ItemChatCardHelper` and the old accessor will be removed with v1.2.0
+- Started deprecation of `ArmorLocation` enum. This enum has been moved to `CONFIG.SWADE.CONST.ARMOR_LOCATIONS` and the old enum will be removed with v1.2.0
+- Started deprecation of `TemplatePreset` enum. This enum has been moved to `CONFIG.SWADE.CONST.TEMPLATE_PRESET` and the old enum will be removed with v1.2.0
 
 ### Removed